// src/app/model/editorial-page.ts
export class EditorialPage {
    pageId: number;
    pageType: string;
    pageName: string;
    createdBy: string;
    updatedBy: string;
    position: number;
    deletabled: boolean;
    kidsHomePage: boolean;
    specialRailAllowed: boolean;
    sectionTypes: string[];
    contentTypes: string[];
    platform: string[];
}